+++ lib/drivers/error_codes.h
@@ -30,7 +30,8 @@ namespace Error {
         None                = 0,
         I2C_Timeout         = -100,
         I2C_Address_Error   = -101,
-        I2C_Data_ACK_Error  = -102
+        I2C_Data_ACK_Error  = -102,
+        I2C_BUS_Busy_Error  = -103
     };
 };