+++ mocking/hardware/vendors.h
@@ -26,7 +26,7 @@
     #define STM32F4 // => The library fakes the STM32F4 microcontroller family
     // === Clock defines ===
     #define F_CPU 40000000
-    #define F_APB 20000000
+    #define F_I2C 20000000
 
     // === Fake peripherals ===
     #include "../stm32/clock_stm32_fake.h"